exercicio6_rag_normas/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0
boto3>=1.34.0
botocore>=1.34.0
scikit-learn>=1.0.0
tf-keras>=2.16.0
